🌳 Refactor: lib 폴더로 관리(hooks, recoil, api etc)
