+++ src/pages/index.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 
 import Content from '@/components/layout/Content';
 import Layout from '@/components/layout/Layout';
-import Sidebar from '@/components/layout/Sidebar';
 
 export default function HomePage() {
   return (
